+++ release-notes.md
@@ -5,6 +5,7 @@ See [log](https://github.com/CLxS/SMAPI/compare/stable...develop).
 
 For players:
   * You can now run most mods on any platform (e.g. run Windows mods on Linux/Mac).
+  * Fixed the normal uninstaller not removing files added by the 'SMAPI for developers' installer.
 
 ## 1.2
 See [log](https://github.com/CLxS/SMAPI/compare/1.1.1...1.2).